update nodes: Timers and Flow
in the Flow section
nodes added:
Save the game
Delete Save slot
If SaveGame exists...

in the Timers section:
added the Create Timer with event set node.
The Start timer node now has a flag to reset the remaining time at the end of the timer.
In the fields where seconds are entered, you can use click + shift to increment or decrement by 1 second and mouse wheel to increment or decrement by 0.1 second